v3

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: jesssanc <jesssanc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/09 13:33:25 by jesssanc          #+#    #+#             */
-/*   Updated: 2024/12/11 13:42:36 by jesssanc         ###   ########.fr       */
+/*   Updated: 2024/12/12 09:05:17 by jesssanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int	ft_printfchar(char c);
 int	ft_printstr(char *s);
 int	ft_printnbr(int n);
 int	ft_printunnbr(unsigned int n);
-int	ft_printhex_low(unsigned int num);
-int	ft_printhex_up(unsigned int num);
+int	ft_printhex_low(unsigned long num);
+int	ft_printhex_up(unsigned long num);
 
 #endif
